Fixed routing issues & renamed React components - Removed history.js, and references to it in app.js & filter.js - Moved animal size & type selectors into case-description - Renamed upload folder to case-upload
