+++ client/src/components/case-upload/dropzone_field.js
