+++ client/src/components/case-upload/image_preview.js
