+++ client/src/components/case-upload/placeholder.js
