+++ client/src/components/case-upload/upload.scss
